Cosealious: add extensible 'SingleResource' API
Summary:
I've created a class that can be used by the `resource` component to load store the resource data. It can be easily extended to have helper funcitons for a
specific purpose, like sending a POST to a related collection, etc.

Test Plan: The diff for T955 should work when used with Sealious with this here patch

Reviewers: #reviewers, piotr-ptaszynski

Reviewed By: #reviewers, piotr-ptaszynski

Subscribers: piotr-ptaszynski

Differential Revision: https://hub.sealcode.org/D303